requirements.txt
gradio
google-auth[pyopenssl]